requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
appnope==0.1.4
argcomplete==3.1.6
asgiref==3.7.2
astrapy==1.1.0
asttokens==3.0.0
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.2.0
blinker==1.7.0
build==1.2.2.post1
cachelib==0.9.0
cachetools==5.5.2
cassandra-driver==3.29.1
cassio==0.1.7
cattrs==23.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
docutils==0.20.1
durationpy==0.9
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.2.0
fastapi==0.115.0
filelock==3.17.0
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
frozenlist==1.4.0
fsspec==2025.3.0
geomet==0.2.1.post1
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==32.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.6
langsmith==0.3.12
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
neo4j==5.20.0
nest-asyncio==1.6.0
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.76.0
openpyxl==3.1.5
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
ordered-set==4.1.0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.0.3
parso==0.8.4
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.3.0
posthog==3.19.1
prison==0.2.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.17.0
PyJWT==2.8.0
pymongo==4.7.2
PyMySQL==1.1.0
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyspark==3.5.5
pyTelegramBotAPI==4.14.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.2.1
redis==5.0.4
referencing==0.31.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.14.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
telebot==0.0.5
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==8.10.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.0.4
uuid6==2024.1.12
uuid_utils==0.7.0
uvicorn==0.31.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
zstandard==0.23.0
